// storeLocatorAngular/src/app/model/cadastro.ts
export class CadastroModel {
  id: number;
  nome: string;
  cpfCnpj: number;
  telefone: number;
  email: string;
  senha: string;
  senhaConfirm: string;
  endereco: string;
  numero: number;
  cep: number;
  bairro: string;
  cidade: string;
  estado: string;
  horaAbertura: string;
  horaFechamento: string;
  categoria: string;
  entrega: string;
}